Add interface and mock for IFileLogsProducer to support logging functionality
